Lint (#3)

+++ protocol/src/time.rs
@@ -6,7 +6,8 @@ pub fn now_utc() -> OffsetDateTime {
 }
 
 #[cfg(target_arch = "wasm32")]
+#[expect(clippy::arithmetic_side_effects, reason = "we know its timestamp")]
 pub fn now_utc() -> OffsetDateTime {
-  let millis = js_sys::Date::now() as i64;
-  OffsetDateTime::from_unix_timestamp_nanos((millis * 1_000_000) as i128).expect("valid timestamp")
+  let millis = js_sys::Date::now() as i128;
+  OffsetDateTime::from_unix_timestamp_nanos(millis * 1_000_000).expect("valid timestamp")
 }
